requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
cmake==3.26.0
easyocr==1.6.2
filelock==3.10.0
idna==3.4
imageio==2.26.0
Jinja2==3.1.2
lazy_loader==0.1
lit==15.0.7
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.0
ninja==1.11.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python-headless==4.5.4.60
packaging==23.0
Pillow==9.4.0
pyclipper==1.3.0.post4
python-bidi==0.4.2
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
scikit-image==0.20.0
scipy==1.10.1
shapely==2.0.1
six==1.16.0
sympy==1.11.1
tifffile==2023.3.15
torch==2.0.0
torchvision==0.15.1
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
